Implement deposit(), withdraw(), and transfer()

+++ src/main/java/com/revature/transactions/Depositable.java
@@ -2,7 +2,9 @@ package com.revature.transactions;
 
 import com.revature.exceptions.InvalidAmountException;
 
+import java.math.BigDecimal;
+
 public interface Depositable
 {
-    public void deposit(double amount) throws InvalidAmountException;
+    public void deposit(BigDecimal amount) throws InvalidAmountException;
 }
