+++ src/main/java/com/revature/transactions/Withdrawable.java
@@ -2,7 +2,9 @@ package com.revature.transactions;
 
 import com.revature.exceptions.InvalidAmountException;
 
+import java.math.BigDecimal;
+
 public interface Withdrawable
 {
-    public void withdraw(double amount) throws InvalidAmountException;
+    public void withdraw(BigDecimal amount) throws InvalidAmountException;
 }
